sim: Split peripherals into unique items
