+++ sim/infra/runtime/c/src/can.c
@@ -5,49 +5,6 @@
 #include "CAN/CAN.h"
 #include "HW.h"
 #include "HW_can.h"
-#include "HW_flash.h"
-
-typedef enum
-{
-    RIG_HW_UART_PORT_UNUSED = 0U,
-} HW_UART_port_E;
-
-RTOS_swiHandle_T* SWI_create(RTOS_swiPri_E priority, RTOS_swiFn_t handler)
-{
-    if ((priority >= RTOS_SWI_PRI_COUNT) || (rig_runtime_swi_count[priority] >= RTOS_SWI_MAX_PER_PRI))
-    {
-        return NULL;
-    }
-
-    const uint8_t   index    = rig_runtime_swi_count[priority]++;
-    RTOS_swiHandle_T* handle = &rig_runtime_swi_handles[priority][index];
-    handle->handler  = handler;
-    handle->priority = priority;
-    handle->event    = 1UL << index;
-    return handle;
-}
-
-void SWI_invoke(RTOS_swiHandle_T* handle)
-{
-    if ((handle != NULL) && (handle->handler != NULL))
-    {
-        handle->handler();
-    }
-}
-
-bool SWI_invokeFromISR(RTOS_swiHandle_T* handle)
-{
-    SWI_invoke(handle);
-    return true;
-}
-
-void SWI_disable(void)
-{
-}
-
-void SWI_enable(void)
-{
-}
 
 HW_StatusTypeDef_E HW_CAN_start(CAN_bus_E bus)
 {
@@ -131,47 +88,4 @@ bool HW_CAN_getRxMessage(CAN_bus_E bus, CAN_RxFifo_E rxFifo, CAN_RxMessage_T* rx
     rx->timestamp        = (uint16_t)(rig_runtime.time_ns / 1000000ULL);
     rx->filterMatchIndex = 0U;
     return true;
-}
-
-HW_StatusTypeDef_E HW_UART_startDMARX(HW_UART_port_E port, uint32_t* data, uint32_t size)
-{
-    (void)port;
-    (void)data;
-    (void)size;
-    return HW_OK;
-}
-
-HW_StatusTypeDef_E HW_UART_stopDMA(HW_UART_port_E port)
-{
-    (void)port;
-    return HW_OK;
-}
-
-uint32_t FLASH_getPageSize(void)
-{
-    return 1024U;
-}
-
-bool FLASH_erasePages(uint32_t pageAddr, uint16_t pages)
-{
-    (void)pageAddr;
-    (void)pages;
-    return true;
-}
-
-bool FLASH_writeHalfwords(uint32_t addr, uint16_t* data, uint16_t dataLen)
-{
-    (void)addr;
-    (void)data;
-    (void)dataLen;
-    return true;
-}
-
-bool HW_mcuShuttingDown(void)
-{
-    return false;
-}
-
-void HW_systemHardReset(void)
-{
 }
