+++ sim/infra/runtime/c/src/system.c
@@ -1,0 +1,10 @@
+#include "rig_runtime.h"
+
+bool HW_mcuShuttingDown(void)
+{
+    return false;
+}
+
+void HW_systemHardReset(void)
+{
+}
