+++ sim/infra/runtime/c/src/uart.c
@@ -1,0 +1,20 @@
+#include "rig_runtime.h"
+
+typedef enum
+{
+    RIG_HW_UART_PORT_UNUSED = 0U,
+} HW_UART_port_E;
+
+HW_StatusTypeDef_E HW_UART_startDMARX(HW_UART_port_E port, uint32_t* data, uint32_t size)
+{
+    (void)port;
+    (void)data;
+    (void)size;
+    return HW_OK;
+}
+
+HW_StatusTypeDef_E HW_UART_stopDMA(HW_UART_port_E port)
+{
+    (void)port;
+    return HW_OK;
+}
